requirements.txt
python-stdnum>=1.5
six>=1.10.0
